refactor OptimizationProblem into MathematicalProgram
because it's a better name now that we are including feasibility programs like dReal and LCP.
because it's potentially confusing to have both class names kicking around
and because there was no advantage that I could see to having the simple boolean logic in MathematicalProgram be separate from OptimizationProblem:
  - hiding the enum could be useful, but that would need to change anyhow when we have the solver interfaces resposible for declaring their own capabilities
  - there is already an observable downside to having them separate -- the dispatch logic is already lagging behind the optimization problem types. plus it's an unnecessary profilieration of AddObjective/Constraint methods.
