+++ drake/solvers/constraint.h
@@ -167,7 +167,7 @@ class SemidefiniteConstraint : public Constraint {
  *
  * The Polynomial class uses a different variable naming scheme; thus the
  * caller must provide a list of Polynomial::VarType variables that correspond
- * to the members of the OptimizationProblem::Binding (the individual scalar
+ * to the members of the MathematicalProgram::Binding (the individual scalar
  * elements of the given VariableList).
  */
 class PolynomialConstraint : public Constraint {
