+++ drake/solvers/dreal_solver.cc
@@ -4,7 +4,6 @@
 #include "dreal/dreal.h"
 
 #include "drake/solvers/mathematical_program.h"
-#include "drake/solvers/optimization.h"
 
 namespace drake {
 namespace solvers {
@@ -13,7 +12,7 @@ bool DrealSolver::available() const {
   return false;
 }
 
-SolutionResult DrealSolver::Solve(OptimizationProblem& prog) const {
+SolutionResult DrealSolver::Solve(MathematicalProgram& prog) const {
   throw std::runtime_error("not implemented yet");
 }
 }  // namespace solvers
