+++ drake/solvers/dreal_solver.h
@@ -2,7 +2,6 @@
 
 #include "drake/drakeOptimization_export.h"
 #include "drake/solvers/mathematical_program.h"
-#include "drake/solvers/optimization.h"
 
 namespace drake {
 namespace solvers {
@@ -13,7 +12,7 @@ class DRAKEOPTIMIZATION_EXPORT DrealSolver
   // This solver is implemented in various pieces depending on if
   // Dreal was available during compilation.
   bool available() const override;
-  SolutionResult Solve(OptimizationProblem& prog) const override;
+  SolutionResult Solve(MathematicalProgram& prog) const override;
 };
 
 }  // namespace solvers
