+++ drake/solvers/equality_constrained_qp_solver.h
@@ -11,7 +11,7 @@ class DRAKEOPTIMIZATION_EXPORT EqualityConstrainedQPSolver
     : public MathematicalProgramSolverInterface {
  public:
   bool available() const override;
-  SolutionResult Solve(OptimizationProblem& prog) const override;
+  SolutionResult Solve(MathematicalProgram& prog) const override;
 };
 
 }  // namespace solvers
