+++ drake/solvers/mathematical_program.cc
@@ -5,7 +5,6 @@
 #include "linear_system_solver.h"
 #include "moby_lcp_solver.h"
 #include "nlopt_solver.h"
-#include "optimization.h"
 #include "snopt_solver.h"
 
 namespace drake {
@@ -13,22 +12,6 @@ namespace solvers {
 
 namespace {
 
-enum ProblemAttributes {
-  kNoCapabilities = 0,
-  kError = 1 << 0,  ///< Do not use, to avoid & vs. && typos.
-  kGenericCost                     = 1 << 1,
-  kGenericConstraint               = 1 << 2,
-  kQuadraticCost                   = 1 << 3,
-  kQuadraticConstraint             = 1 << 4,
-  kLinearCost                      = 1 << 5,
-  kLinearConstraint                = 1 << 6,
-  kLinearEqualityConstraint        = 1 << 7,
-  kLinearComplementarityConstraint = 1 << 8
-};
-typedef uint32_t AttributesSet;
-
-// TODO(ggould-tri) Refactor these capability advertisements into the
-// solver wrappers themselves.
 
 // Solver for simple linear systems of equalities
 AttributesSet kLinearSystemSolverCapabilities = kLinearEqualityConstraint;
@@ -60,8 +43,19 @@ bool is_satisfied(AttributesSet required, AttributesSet available) {
 }  // anon namespace
 
 
+enum {
+  INITIAL_VARIABLE_ALLOCATION_NUM = 100
+};  // not const static int because the VectorXd constructor takes a reference
+// to int so it is odr-used (see
+// https://gcc.gnu.org/wiki/VerboseDiagnostics#missing_static_const_definition)
+
+
 MathematicalProgram::MathematicalProgram()
-    : required_capabilities_(kNoCapabilities),
+    : num_vars_(0),
+      x_initial_guess_(
+                  static_cast<Eigen::Index>(INITIAL_VARIABLE_ALLOCATION_NUM)),
+      solver_result_(0),
+      required_capabilities_(kNoCapabilities),
       ipopt_solver_(new IpoptSolver()),
       nlopt_solver_(new NloptSolver()),
       snopt_solver_(new SnoptSolver()),
@@ -69,32 +63,8 @@ MathematicalProgram::MathematicalProgram()
       linear_system_solver_(new LinearSystemSolver()),
       equality_constrained_qp_solver_(new EqualityConstrainedQPSolver()) {}
 
-void MathematicalProgram::AddGenericCost() {
-  required_capabilities_ |= kGenericCost;
-}
-void MathematicalProgram::AddGenericConstraint() {
-  required_capabilities_ |= kGenericConstraint;
-}
-void MathematicalProgram::AddQuadraticCost() {
-  required_capabilities_ |= kQuadraticCost;
-}
-void MathematicalProgram::AddQuadraticConstraint() {
-  required_capabilities_ |= kQuadraticConstraint;
-}
-void MathematicalProgram::AddLinearCost() {
-  required_capabilities_ |= kLinearCost;
-}
-void MathematicalProgram::AddLinearConstraint() {
-  required_capabilities_ |= kLinearConstraint;
-}
-void MathematicalProgram::AddLinearEqualityConstraint() {
-  required_capabilities_ |= kLinearEqualityConstraint;
-}
-void MathematicalProgram::AddLinearComplementarityConstraint() {
-  required_capabilities_ |= kLinearComplementarityConstraint;
-}
 
-SolutionResult MathematicalProgram::Solve(OptimizationProblem& prog) const {
+SolutionResult MathematicalProgram::Solve()  {
   // This implementation is simply copypasta for now; in the future we will
   // want to tweak the order of preference of solvers based on the types of
   // constraints present.
@@ -104,23 +74,23 @@ SolutionResult MathematicalProgram::Solve(OptimizationProblem& prog) const {
     // TODO(ggould-tri) Also allow quadratic objectives whose matrix is
     // Identity: This is the objective function the solver uses anyway when
     // underconstrainted, and is fairly common in real-world problems.
-    return linear_system_solver_->Solve(prog);
+    return linear_system_solver_->Solve(*this);
   } else if (is_satisfied(required_capabilities_,
                           kEqualityConstrainedQPCapabilities) &&
             equality_constrained_qp_solver_->available()) {
-    return equality_constrained_qp_solver_->Solve(prog);
+    return equality_constrained_qp_solver_->Solve(*this);
   } else if (is_satisfied(required_capabilities_, kMobyLcpCapabilities) &&
              moby_lcp_solver_->available()) {
-    return moby_lcp_solver_->Solve(prog);
+    return moby_lcp_solver_->Solve(*this);
   } else if (is_satisfied(required_capabilities_, kGenericSolverCapabilities) &&
              snopt_solver_->available()) {
-    return snopt_solver_->Solve(prog);
+    return snopt_solver_->Solve(*this);
   } else if (is_satisfied(required_capabilities_, kGenericSolverCapabilities) &&
              ipopt_solver_->available()) {
-    return ipopt_solver_->Solve(prog);
+    return ipopt_solver_->Solve(*this);
   } else if (is_satisfied(required_capabilities_, kGenericSolverCapabilities) &&
              nlopt_solver_->available()) {
-    return nlopt_solver_->Solve(prog);
+    return nlopt_solver_->Solve(*this);
   } else {
     throw std::runtime_error(
         "MathematicalProgram::Solve: "
