+++ drake/solvers/mosek_solver.cc
@@ -6,16 +6,12 @@
 
 #include <Eigen/Core>
 
-#include "drake/solvers/optimization.h"
-#include "drake/solvers/mathematical_program.h"
-#include "drake/solvers/solution_result.h"
-
 namespace drake {
 namespace solvers {
 
 bool MosekSolver::available() const { return true; }
 
-SolutionResult MosekSolver::Solve(OptimizationProblem& prog) const {
+SolutionResult MosekSolver::Solve(MathematicalProgram& prog) const {
   if (!prog.GetSolverOptionsStr("Mosek").empty()) {
     if (prog.GetSolverOptionsStr("Mosek").at("problemtype").find("linear")
             != std::string::npos ||
