+++ drake/solvers/no_ipopt.cc
@@ -10,7 +10,7 @@ bool IpoptSolver::available() const {
   return false;
 }
 
-SolutionResult IpoptSolver::Solve(OptimizationProblem &prog) const {
+SolutionResult IpoptSolver::Solve(MathematicalProgram &prog) const {
   throw std::runtime_error(
       "The IPOPT bindings were not compiled.  You'll need to use a different "
       "solver.");
