+++ drake/solvers/no_mosek.cc
@@ -10,7 +10,7 @@ bool MosekSolver::available() const {
   return false;
 }
 
-SolutionResult MosekSolver::Solve(OptimizationProblem &prog) const {
+SolutionResult MosekSolver::Solve(const MathematicalProgram &prog) const {
   throw std::runtime_error(
       "Mosek is not installed in your build. You'll need to use a different "
       "solver.");
