+++ drake/solvers/no_nlopt.cc
@@ -10,7 +10,7 @@ bool NloptSolver::available() const {
   return false;
 }
 
-SolutionResult NloptSolver::Solve(OptimizationProblem &prog) const {
+SolutionResult NloptSolver::Solve(const MathematicalProgram &prog) const {
   throw std::runtime_error(
       "The Nlopt bindings were not compiled.  You'll need to use a different "
       "solver.");
