+++ drake/solvers/no_snopt.cc
@@ -10,7 +10,7 @@ bool SnoptSolver::available() const {
   return false;
 }
 
-SolutionResult SnoptSolver::Solve(OptimizationProblem &prog) const {
+SolutionResult SnoptSolver::Solve(const MathematicalProgram &prog) const {
   throw std::runtime_error(
       "The SNOPT bindings were not compiled.  You'll need to use a different "
       "solver.");
