+++ drake/systems/plants/inverseKinBackend.cpp
@@ -11,7 +11,7 @@
 #include "drake/math/autodiff.h"
 #include "drake/math/autodiff_gradient.h"
 #include "drake/math/gradient.h"
-#include "drake/solvers/optimization.h"
+#include "drake/solvers/mathematical_program.h"
 #include "drake/systems/plants/constraint/RigidBodyConstraint.h"
 #include "drake/systems/plants/ConstraintWrappers.h"
 #include "drake/systems/plants/IKoptions.h"
@@ -27,7 +27,7 @@ using drake::math::autoDiffToGradientMatrix;
 using drake::math::autoDiffToValueMatrix;
 using drake::solvers::Constraint;
 using drake::solvers::DecisionVariableView;
-using drake::solvers::OptimizationProblem;
+using drake::solvers::MathematicalProgram;
 using drake::solvers::SolutionResult;
 
 /// NOTE: The contents of this class are for the most part direct ports of
@@ -39,7 +39,7 @@ namespace drake {
 namespace systems {
 namespace plants {
 
-int GetIKSolverInfo(const OptimizationProblem& prog, SolutionResult result) {
+int GetIKSolverInfo(const MathematicalProgram& prog, SolutionResult result) {
   std::string solver_name;
   int solver_result = 0;
   prog.GetSolverResult(&solver_name, &solver_result);
@@ -69,7 +69,7 @@ int GetIKSolverInfo(const OptimizationProblem& prog, SolutionResult result) {
 }
 
 void SetIKSolverOptions(const IKoptions& ikoptions,
-                        drake::solvers::OptimizationProblem* prog) {
+                        drake::solvers::MathematicalProgram* prog) {
   prog->SetSolverOption("SNOPT", "Derivative option", 1);
   prog->SetSolverOption("SNOPT", "Major optimality tolerance",
                         ikoptions.getMajorOptimalityTolerance());
@@ -86,7 +86,7 @@ void SetIKSolverOptions(const IKoptions& ikoptions,
 void AddSingleTimeLinearPostureConstraint(
     const double *t, const RigidBodyConstraint* constraint, int nq,
     const drake::solvers::DecisionVariableView& vars,
-    OptimizationProblem* prog) {
+    MathematicalProgram* prog) {
   DRAKE_ASSERT(
       constraint->getCategory() ==
       RigidBodyConstraint::SingleTimeLinearPostureConstraintCategory);
@@ -125,7 +125,7 @@ void AddQuasiStaticConstraint(
     const double *t, const RigidBodyConstraint* constraint,
     KinematicsCacheHelper<double>* kin_helper,
     const drake::solvers::DecisionVariableView& vars,
-    drake::solvers::OptimizationProblem* prog) {
+    drake::solvers::MathematicalProgram* prog) {
   DRAKE_ASSERT(constraint->getCategory() ==
                RigidBodyConstraint::QuasiStaticConstraintCategory);
 
@@ -210,12 +210,12 @@ void inverseKinBackend(
   KinematicsCacheHelper<double> kin_helper(model->bodies);
 
   // TODO(sam.creasey) I really don't like rebuilding the
-  // OptimizationProblem for every timestep, but it's not possible to
+  // MathematicalProgram for every timestep, but it's not possible to
   // enable/disable (or even remove) a constraint from an
-  // OptimizationProblem between calls to Solve() currently, so
+  // MathematicalProgram between calls to Solve() currently, so
   // there's not actually another way.
   for (int t_index = 0; t_index < nT; t_index++) {
-    OptimizationProblem prog;
+    MathematicalProgram prog;
     SetIKSolverOptions(ikoptions, &prog);
 
     DecisionVariableView vars =
